pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gcodeui"
version = "0.1.0"
description = "A G-code sender UI for CNC machines and 3D printers"
readme = "README.md"
authors = [{ name = "Shachar Fleishman" }]
license = { file = "LICENSE" }
requires-python = ">=3.7"
keywords = ["gcode", "cnc", "3d-printing", "tkinter", "serial"]
dependencies = [
    "tk",
    "pyserial",
    "pyyaml",
    "addict",
    "structlog",
    "black",
]

[project.urls]
Homepage = "https://github.com/shachar/gcodeui"

[project.scripts]
gcodeui = "gcodeui:main"

[tool.setuptools]
py-modules = ["gcodeui"]
include-package-data = true

[tool.setuptools.package-data]
"" = ["config.yaml"]
